+++ payment-service/routes/api.js
@@ -1,8 +1,8 @@
 import express from 'express';
-import makePayment from './payment'
+import makePayment from './paymentRoutes.js'; 
 
-const api = express.Router();
+const api = express.Router(); 
 
 api.use('/payments', makePayment);
 
-export default api;
+export default api;
