--- payment-service/routes/payment.js
@@ -1,9 +1,0 @@
-import express from 'express';
-import { createPayment } from '../controllers/createPayment.js'; // Ensure this matches exactly
-import authMiddleware from '../middlewares/authMiddleware.js';
-
-const makePayment = express.Router();
-
-makePayment.post('/payments', authMiddleware, createPayment); 
-
-export default makePayment;
